requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
huggingface_hub